リリースノートに更新機能を実装した

+++ app/Http/Controllers/ReleaseNoteController.php
@@ -61,25 +61,35 @@ class ReleaseNoteController extends Controller
 
     /**
      * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(ReleaseNote $release_note)
     {
-        //
+        if (Auth::user()->email !== "sunasunayaka1218@gmail.com") {
+            abort(404);
+        }
+
+        return view('release-note.edit')->with('release_note', $release_note);
     }
 
     /**
      * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, ReleaseNote $release_note)
     {
-        //
+        if (Auth::user()->email !== "sunasunayaka1218@gmail.com") {
+            abort(404);
+        }
+
+        $release_note_data = $request->validate([
+            'title' => ['required', 'min:3'],
+            'detail' => ['required']
+        ]);
+
+        $release_note->title = $release_note_data['title'];
+        $release_note->detail = $release_note_data['detail'];
+        $release_note->save();
+
+        return redirect()->route('release-note.show', $release_note->id);
     }
 
     /**
@@ -87,6 +97,10 @@ class ReleaseNoteController extends Controller
      */
     public function destroy(ReleaseNote $release_note)
     {
+        if (Auth::user()->email !== "sunasunayaka1218@gmail.com") {
+            abort(404);
+        }
+
         $release_note->delete();
 
         return redirect()->route('release-note.index');
